// apps/web/src/lib/service-clients/service-properties/generated/schemas/createPropertyDefinitionRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * properties_service
 * OpenAPI spec version: 0.1.0
 */

import type { CreatePropertyScope } from './createPropertyScope';
import type { PropertyDataType } from './propertyDataType';

/**
 * Request to create a new property definition.
 */
export interface CreatePropertyDefinitionRequest {
  data_type: PropertyDataType;
  display_name: string;
  scope: CreatePropertyScope;
}