// apps/web/src/lib/service-clients/service-properties/generated/schemas/entityPropertiesResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * properties_service
 * OpenAPI spec version: 0.1.0
 */
import type { EntityPropertyWithDefinition } from './entityPropertyWithDefinition';

/**
 * Response for document/entity properties endpoint.
 */
export interface EntityPropertiesResponse {
  entity_id: string;
  properties: EntityPropertyWithDefinition[];
}