// apps/web/src/lib/service-clients/service-properties/generated/schemas/listPropertiesParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * properties_service
 * OpenAPI spec version: 0.1.0
 */

import type { EntityType } from './entityType';
import type { PropertyScope } from './propertyScope';

export type ListPropertiesParams = {
  /**
   * Filter by scope: 'user', 'team', 'system', or 'all'
   */
  scope: PropertyScope;
  /**
   * Whether to include property options in the response
   */
  include_options?: boolean;
  /**
   * Filter properties applicable to a specific entity type
   */
  for_entity_type?: EntityType;
};